Move over validate media
